feat: add actions dropdown menu
